refactor(shared): move settings mutex to $/shared, dedup PluginLike
globalSettingsMutex lived in command-permissions, which made that
feature the de-facto home of process-wide data.json infrastructure and
gave it 12+ inbound edges. Move settingsLock.ts (and its test) to
$/shared; command-permissions re-exports it as a forwarding shim so the
six barrel consumers stay unchanged, only three deep-path imports move.
Add $/shared/types.ts with PluginReadLike/PluginDataLike and route the
local PluginLike redeclarations through it, fixing the mcp-client-config
to adaptive-tool-loading cross-feature type import. Behavior-preserving.
