+++ packages/obsidian-plugin/src/features/command-permissions/index.ts
@@ -8,11 +8,14 @@ export {
   type CommandPermissionModalOptions,
   type ModalDecision,
 } from "./services/commandPermissionModal";
+// Forwarding shim: the settings mutex now lives in $/shared (it is
+// process-wide infrastructure, not a permissions concern). Re-exported
+// here so existing barrel consumers keep working unchanged.
 export {
   createMutex,
   globalSettingsMutex,
   type Mutex,
-} from "./services/settingsLock";
+} from "$/shared/settingsLock";
 export type { CommandAuditEntry } from "./types";
 export {
   AUDIT_LOG_MAX_ENTRIES,
