+++ packages/obsidian-plugin/src/shared/settingsLock.test.ts
