+++ packages/obsidian-plugin/src/shared/settingsLock.ts
